Move related type definitions back to Camera for doc generation
